Backend/multiplayer_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.0
requests==2.31.0
PyJWT==2.8.0
bcrypt==4.1.2
email-validator==2.1.0